requirements.txt
streamlit==1.28.1
sentence-transformers==2.2.2
chromadb==0.4.18
rank-bm25==0.2.2
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
openai==0.28.1
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
transformers==4.35.0
markdown==3.5.1